Refactored, added documentation.
