+++ src/main/java/io/anserini/rerank/lib/TiebreakerReranker.java
@@ -1,5 +1,8 @@
-package io.anserini.rerank;
+package io.anserini.rerank.lib;
 
+import io.anserini.rerank.Reranker;
+import io.anserini.rerank.RerankerContext;
+import io.anserini.rerank.ScoredDocuments;
 import org.apache.lucene.document.Document;
 
 import java.util.SortedSet;
@@ -8,10 +11,20 @@ import java.util.TreeSet;
 import static io.anserini.index.generator.LuceneDocumentGenerator.FIELD_ID;
 
 /**
- * Created by jimmylin on 6/19/18.
+ * Reranker that ensures consistent ordering of documents that have the same score. Scoring ties are broken based on the
+ * lexicographic ordering of the collection docid. This is accomplished by rounding original document scores to the
+ * fourth decimal place, and then adding a tiny score perturbation to break scoring ties.
+ *
+ * This is necessary for repeatable runs: due to multi-threaded indexing, documents are added to the index in arbitrary
+ * order, which makes Lucene's internal mechanism for resolving scoring ties non-deterministic across different indexes.
+ *
+ * Note however, that this reranker also is not sufficient for completely repeatable runs due to scoring ties that span
+ * the rank cutoff <i>k</i>. Due to scoring ties, the top <i>k</i> might vary across indexes; there is nothing that this
+ * reranker can do for such cases. The only solution is to retrieve more than top <i>k</i>, break scoring ties, and then
+ * truncate to tope <i>k</i>.
  */
 public class TiebreakerReranker implements Reranker {
-    // Sort by score, break ties by sort order of collection docid.
+    // Sort by score, break ties by lexicographic ordering of collection docid.
     private class Result implements Comparable<Result> {
       public float score;
       public String docid;
